moved full restart test
